changed the changePassword function to use the encrypted private key rather than retrieve it from locksmith. This will allow for re-use of that function when doing key recovery
