+++ unlock-app/src/__tests__/middlewares/providerMiddleware.test.ts
@@ -1,11 +1,13 @@
 import { createAccountAndPasswordEncryptKey } from '@unlock-protocol/unlock-js'
 import providerMiddleware, {
+  changePassword,
   initializeUnlockProvider,
 } from '../../middlewares/providerMiddleware'
 import { SET_PROVIDER, providerReady } from '../../actions/provider'
 import { setError } from '../../actions/error'
 import { FATAL_MISSING_PROVIDER } from '../../errors'
 import { Application, LogIn } from '../../utils/Error'
+
 import {
   GOT_ENCRYPTED_PRIVATE_KEY_PAYLOAD,
   SIGN_USER_DATA,
@@ -245,6 +247,49 @@ describe('provider middleware', () => {
         type: SIGN_PURCHASE_DATA,
         data,
       })
+    })
+  })
+
+  describe('CHANGE_PASSWORD', () => {
+    const encryptedPrivateKey = {
+      version: 3,
+      id: 'edbe0942-593b-4027-8688-07b7d3ec56c5',
+      address: '0272742cbe9b4d4c81cffe8dfc0c33b5fb8893e5',
+      crypto: {
+        ciphertext:
+          '6f2a3ed499a2962cc48e6f7f0a90a0c817c83024cc4878f624ad251fccd0b706',
+        cipherparams: { iv: '69f031944591eed34c4d4f5841d283b0' },
+        cipher: 'aes-128-ctr',
+        kdf: 'scrypt',
+        kdfparams: {
+          dklen: 32,
+          salt:
+            '5ac866336768f9613a505acd18dab463f4d10152ffefba5772125f5807539c36',
+          n: 8192,
+          r: 8,
+          p: 1,
+        },
+        mac: 'cc8efad3b534336ecffc0dbf6f51fd558301873d322edc6cbc1c9398ee0953ec',
+      },
+    }
+    const oldPassword = 'guest'
+
+    it('should dispatch a new user object to be signed', async () => {
+      expect.assertions(1)
+      const dispatch = jest.fn()
+
+      await changePassword({
+        oldPassword,
+        newPassword: 'visitor',
+        passwordEncryptedPrivateKey: encryptedPrivateKey,
+        dispatch,
+      })
+
+      expect(dispatch).toHaveBeenCalledWith(
+        expect.objectContaining({
+          type: SIGN_USER_DATA,
+        })
+      )
     })
   })
 })
